Clean design and remove chat api
