+++ src/main/resources/application.properties
@@ -1,41 +1,6 @@
-# Primary LLM Gate - must be explicitly enabled for LLM functionality
-app.llm.enable=${ENABLE_LLM:false}
-
-# LLM Provider Configuration (configurable: openai, ollama, or none)
-app.llm.provider=${LLM_PROVIDER:none}
-
 # =============================================================================
-# LLM Configuration
-# =============================================================================
-
-# LangChain4J Provider Configuration
-# Always set a valid provider for build-time, but our runtime code controls usage
-quarkus.langchain4j.chat-model.provider=${LLM_PROVIDER:ollama}
-
-# Disable all LangChain4J dev services to prevent automatic container startup
-quarkus.langchain4j.devservices.enabled=false
-
-# -----------------------------------------------------------------------------
-# Ollama Configuration (Primary - Local/Kubernetes)
-# -----------------------------------------------------------------------------
-quarkus.langchain4j.ollama.base-url=${OLLAMA_BASE_URL:http://localhost:11434}
-quarkus.langchain4j.ollama.chat-model.model-id=llama3.2:3b
-quarkus.langchain4j.ollama.chat-model.temperature=0.3
-quarkus.langchain4j.ollama.timeout=120s
-quarkus.langchain4j.ollama.log-requests=true
-quarkus.langchain4j.ollama.log-responses=true
-
-
-# -----------------------------------------------------------------------------
-# OpenAI Configuration (Optional - Cloud)
-# -----------------------------------------------------------------------------
-quarkus.langchain4j.openai.api-key=${OPENAI_API_KEY:not-configured}
-quarkus.langchain4j.openai.chat-model.model-name=gpt-4o-mini
-quarkus.langchain4j.openai.chat-model.temperature=0.3
-quarkus.langchain4j.openai.timeout=60s
-
-
 # MCP Server Configuration
+# =============================================================================
 quarkus.http.cors.enabled=true
 quarkus.http.cors.origins=/.*/
 quarkus.http.cors.methods=GET,POST,OPTIONS
@@ -49,15 +14,10 @@ quarkus.kubernetes-client.namespace=${K8S_NAMESPACE:kafka}
 quarkus.http.port=8080
 
 # Logging Configuration
-quarkus.log.category."io.strimzi.mcp".level=DEBUG
-quarkus.log.category."dev.langchain4j".level=INFO
+quarkus.log.category."io.streamshub.mcp".level=DEBUG
 quarkus.log.console.format=%d{HH:mm:ss} %-5p [%c{2.}] (%t) %s%e%n
 
 # Development Configuration
-%dev.quarkus.langchain4j.openai.log-requests=true
-%dev.quarkus.langchain4j.openai.log-responses=true
-%dev.quarkus.langchain4j.ollama.log-requests=true
-%dev.quarkus.langchain4j.ollama.log-responses=true
 
 # Reduce dev mode noise
 %dev.quarkus.log.category."io.quarkus.deployment.dev.DevModeCommandLineBuilder".level=WARN
